+++ src/lambda/lambdaTreeDataProvider.ts
@@ -11,10 +11,10 @@ import { RegionProvider } from '../shared/regions/regionProvider'
 import { ResourceFetcher } from '../shared/resourceFetcher'
 import { TelemetryNamespace } from '../shared/telemetry/telemetryTypes'
 import { registerCommand } from '../shared/telemetry/telemetryUtils'
-import { AWSCommandTreeNode } from '../shared/treeview/awsCommandTreeNode'
-import { AWSTreeNodeBase } from '../shared/treeview/awsTreeNodeBase'
 import { RefreshableAwsTreeProvider } from '../shared/treeview/awsTreeProvider'
-import { ErrorNode } from '../shared/treeview/errorNode'
+import { AWSCommandTreeNode } from '../shared/treeview/nodes/awsCommandTreeNode'
+import { AWSTreeNodeBase } from '../shared/treeview/nodes/awsTreeNodeBase'
+import { ErrorNode } from '../shared/treeview/nodes/errorNode'
 import { intersection, toMap, updateInPlace } from '../shared/utilities/collectionUtils'
 import { localize } from '../shared/utilities/vsCodeUtils'
 import { deleteCloudFormation } from './commands/deleteCloudFormation'
