+++ src/shared/treeview/nodes/errorNode.ts
