+++ src/shared/treeview/nodes/placeholderNode.ts
